+++ src/main/java/com/expickpay/pickpay/entity/TransactionEntity.java
@@ -6,6 +6,7 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Entity
@@ -27,6 +28,7 @@ public class TransactionEntity {
     @JoinColumn(name = "receiver_id")
     private UserEntity receiver;
 
-    private double amount;
+    private BigDecimal amount;
     private LocalDateTime date;
+
 }
